feat: Simplified global state
There's a lot of kernel state that unfortunately has to be global,
because it's accessed from interrupt handlers, Rust's memory allocation
system, or other context-less locations.

This state must all be initialized in a specific order, and is then
generally safe to pass around as `&static` references. Adding a
`Once`-like `Global` type captures this, and marks such global state for
future improvements.