feat: Start of HAL model
This moves back to a HAL model, primarily so that the kernel can be
split into sub-crates that all access platform-specific implementations
via the HAL.

The kernel itself is still architecture-aware via the `arch` module.
This means that `arch` can export concrete types of HAL implementations
for the rest of the kernel to use. That's pretty unavoidable because
static variables that rely on HAL types can't be generic.